fix(money): store impact figures as DECIMAL, not single-precision float
`policy_premium_impact` and `direct_dollar_impact` were `DataTypes.REAL`, which
Sequelize maps to float4 on Postgres — about seven significant digits — so the
STORED value was wrong, not just arithmetic on it:

    1234567.89  ->  1234567.875          (displays as $1,234,567.88)
      99999.99  ->  99999.9921875
          0.07  ->  0.07000000029802322

SQLite's REAL is a double, so nothing reproduced on the local sqljs path; this
only ever damaged hosted data. The Excel export writes these values out raw, so
it showed in spreadsheets as well as on screen.

Now DECIMAL(14,2). The cost is a type change on the wire: pg returns numeric as
a STRING and Sequelize's Postgres DECIMAL.parse passes it through to preserve
precision, so a naive change would have turned 1250 into "1250.00" in every
admin response. `toMoneyNumber` coerces both back in mapSubmission — the one
boundary every submission payload and socket emit already passes through — so
the JSON contract is identical on both dialects. Null stays null, not 0:
"nobody costed this ticket" and "zero dollars" are different answers and the
impact totals must not conflate them.

Migration is explicit rather than a boot-sync side effect. Production boots with
sync({ alter: true }), which would have performed this ALTER silently on the next
deploy; a column-type change on live data should be an act someone reads the
output of. `npm run migrate:money-columns` is dry-run by default, applies in one
transaction, and reports how many rows carry the float4 damage signature. It
cannot recover precision already lost.

Applied to production: 83 rows, 1 money value (12 -> 12.00), 0 damaged. The fix
lands before the columns fill with real money rather than after.

One correction made in the doing: the initial diagnosis claimed the queue's
impact totals accumulated visible float error. Measured, that was wrong — float64
summation stays under half a cent even at 50,000 rows, so formatCurrency's 2dp
rounding already hid it. The totals now sum in integer cents anyway (exact rather
than exact-after-rounding), but it is defensive and the comments say so.

Not fixed, deliberately: the ISO-strings-in-TEXT timestamp columns. `updated_at`
doubles as the optimistic-concurrency token and is compared as a STRING, so a
native-timestamp conversion would fail the JSON round-trip on precision and 409
every save. Specified for the rebuild in docs/handoff instead.

250 tests (7 new), lint and build green. Verified end to end against a local
instance: 1234567.89 and 0.07 round-trip exactly as JSON numbers, an uncosted
ticket reads null, and the public endpoint still withholds both fields.

+++ server/db/models/index.js
@@ -121,8 +121,27 @@ function defineModels(sequelize) {
     desired_completion_date: { type: DataTypes.TEXT },
     impact_details: { type: DataTypes.TEXT },
     impact_notes: { type: DataTypes.TEXT },
-    policy_premium_impact: { type: DataTypes.REAL },
-    direct_dollar_impact: { type: DataTypes.REAL },
+    // Money. DECIMAL, deliberately, not REAL — and the distinction is not
+    // academic: Sequelize's REAL is single-precision (float4) on Postgres, about
+    // seven significant digits, so the STORED value was wrong:
+    //
+    //     1234567.89  ->  1234567.875          (displays as $1,234,567.88)
+    //       99999.99  ->  99999.9921875
+    //           0.07  ->  0.07000000029802322
+    //
+    // SQLite's REAL is a double, which is why this never reproduced locally and
+    // only ever damaged the hosted data. The export writes these values straight
+    // into a spreadsheet, so the damage was visible there as well as a cent adrift
+    // on screen.
+    //
+    // Cost of the change: `pg` returns `numeric` as a STRING, and Sequelize's
+    // Postgres DECIMAL.parse passes it through unchanged. `mapSubmission`
+    // (helpers/mappers.js) coerces both back to numbers so the JSON contract
+    // stays numeric on every dialect — every submission response and socket
+    // payload goes through that one mapper, so that is the only place it is
+    // needed. Do not read these columns around it.
+    policy_premium_impact: { type: DataTypes.DECIMAL(14, 2) },
+    direct_dollar_impact: { type: DataTypes.DECIMAL(14, 2) },
     policies_affected_count: { type: DataTypes.INTEGER },
     logged_defect: { type: DataTypes.INTEGER, allowNull: false, defaultValue: 0 },
     // Raised by the rep on the submit form: they are blocked now and need a
